Make library compilable for Native Dart #38 #32
