+++ lib/client_io.dart
@@ -1,0 +1,9 @@
+import 'package:pwa/interface.dart';
+
+class Client implements BaseClient {
+  @override
+  Future<PushPermission> getPushPermission({bool subscribeIfNeeded = false}) {}
+
+  @override
+  bool get isSupported => false;
+}
